city_backend/citycompeer/requirements.txt
Package                      Version
---------------------------- ----------
 
decouple                     0.0.7 
Django                       4.1.5
django-cors-headers          4.3.1
djangorestframework          3.14.0  
keras                        2.15.0  
nltk                         3.8.1
numpy                        1.25.1    
tensorflow                   2.15.0  